Struct rdf_rs::reader::input_reader::InputReader

pub struct InputReader<R: Read> { /* fields omitted */ }

Reads input and transforms it to InputChars.

Methods

impl<R: Read> InputReader<R>

fn new(input: R) -> InputReader<R>

Constructor for InputReader.

Examples

use rdf_rs::reader::input_reader::InputReader;

let reader = InputReader::new("_:auto0".as_bytes());

fn peek_next_k_chars(&mut self, k: usize) -> Result<InputChars>

Returns the next k characters but does not consume them.

Examples

use rdf_rs::reader::input_reader::InputReader;
use rdf_rs::reader::input_reader::InputChars;

let mut reader = InputReader::new("_:auto0".as_bytes());
assert_eq!(reader.peek_next_k_chars(2).unwrap().to_vec(), vec![Some('_'), Some(':')]);
assert_eq!(reader.peek_next_k_chars(2).unwrap().to_vec(), vec![Some('_'), Some(':')]);

Failures

End of input reached.

fn peek_next_char(&mut self) -> Result<Option<char>>

Returns the next character but does not consume it.

Examples

use rdf_rs::reader::input_reader::InputReader;
use rdf_rs::reader::input_reader::InputChars;

let mut reader = InputReader::new("_:auto0".as_bytes());
assert_eq!(reader.peek_next_char().unwrap(), Some('_'));
assert_eq!(reader.peek_next_char().unwrap(), Some('_'));

Failures

End of input reached.

fn peek_next_char_discard_leading_spaces(&mut self) -> Result<Option<char>>

Returns the next character that is not a whitespace but does not consume it.

Examples

use rdf_rs::reader::input_reader::InputReader;
use rdf_rs::reader::input_reader::InputChars;

let mut reader = InputReader::new(" \n _:auto0".as_bytes());
assert_eq!(reader.peek_next_char_discard_leading_spaces().unwrap(), Some('_'));
assert_eq!(reader.peek_next_char_discard_leading_spaces().unwrap(), Some('_'));

Failures

End of input reached.

fn get_next_char(&mut self) -> Result<Option<char>>

Returns the next character of an input source.

Examples

use rdf_rs::reader::input_reader::InputReader;

let mut input = "Hello World!".as_bytes();
let mut input_reader = InputReader::new(input);

assert_eq!(Some('H'), input_reader.get_next_char().unwrap());
assert_eq!(Some('e'), input_reader.get_next_char().unwrap());

Failures

End of input reached.

fn get_next_k_chars(&mut self, k: usize) -> Result<InputChars>

Returns the next k characters of an input source and consumes them.

Examples

use rdf_rs::reader::input_reader::InputReader;
use rdf_rs::reader::input_reader::InputChars;

let mut reader = InputReader::new("_:auto0".as_bytes());
assert_eq!(reader.get_next_k_chars(2).unwrap().to_vec(), vec![Some('_'), Some(':')]);
assert_eq!(reader.get_next_k_chars(2).unwrap().to_vec(), vec![Some('a'), Some('u')]);

Failures

End of input reached.

fn get_next_char_discard_leading_spaces(&mut self) -> Result<Option<char>>

Returns the next character of an input source that is not a whitespace.

Examples

use rdf_rs::reader::input_reader::InputReader;

let mut input = "H   ello World!".as_bytes();
let mut input_reader = InputReader::new(input);

assert_eq!(Some('H'), input_reader.get_next_char_discard_leading_spaces().unwrap());
assert_eq!(Some('e'), input_reader.get_next_char_discard_leading_spaces().unwrap());

Failures

End of input reached.

fn peek_until<F: Fn(char) -> bool>(&mut self,
delimiter: F)
-> Result<InputChars>

Returns all characters of a input source until a certain delimiter occurs but does not consume them.

The delimiter itself is skipped.

Examples

use rdf_rs::reader::input_reader::InputReader;

let mut input = "Hello World!".as_bytes();
let mut input_reader = InputReader::new(input);

assert_eq!("Hello".to_string(), input_reader.peek_until(|c| c == ' ').unwrap().to_string());
assert_eq!("Hello".to_string(), input_reader.peek_until(|c| c == ' ').unwrap().to_string());

Failures

End of input reached.

fn peek_until_discard_leading_spaces<F: Fn(char) -> bool>(&mut self,
delimiter: F)
-> Result<InputChars>

Returns all characters without consuming them of a input source until a certain delimiter occurs and removes leading whitespaces.

The delimiter itself is skipped.

Examples

use rdf_rs::reader::input_reader::InputReader;

let mut input = "   Hello    World!".as_bytes();
let mut input_reader = InputReader::new(input);

assert_eq!("Hello".to_string(), input_reader.peek_until_discard_leading_spaces(|c| c == ' ').unwrap().to_string());
assert_eq!("Hello".to_string(), input_reader.peek_until_discard_leading_spaces(|c| c == ' ').unwrap().to_string());

Failures

End of input reached.

fn get_until<F: Fn(char) -> bool>(&mut self, delimiter: F) -> Result<InputChars>

Returns all characters of a input source until a certain delimiter occurs.

The delimiter itself is skipped.

Examples

use rdf_rs::reader::input_reader::InputReader;

let mut input = "Hello World!".as_bytes();
let mut input_reader = InputReader::new(input);

assert_eq!("Hello".to_string(), input_reader.get_until(|c| c == ' ').unwrap().to_string());
assert_eq!(" World".to_string(), input_reader.get_until(|c| c == '!').unwrap().to_string());

Failures

End of input reached.

fn get_until_discard_leading_spaces<F: Fn(char) -> bool>(&mut self,
delimiter: F)
-> Result<InputChars>

Returns all characters of a input source until a certain delimiter occurs and removes leading whitespaces.

The delimiter itself is skipped.

Examples

use rdf_rs::reader::input_reader::InputReader;

let mut input = "Hello    World!".as_bytes();
let mut input_reader = InputReader::new(input);

assert_eq!("Hello".to_string(), input_reader.get_until_discard_leading_spaces(|c| c == ' ').unwrap().to_string());
assert_eq!("World".to_string(), input_reader.get_until_discard_leading_spaces(|c| c == '!').unwrap().to_string());

Failures

End of input reached.