Struct rdf_rs::triple::TripleStore

pub struct TripleStore { /* fields omitted */ }

Storage for triples.

Methods

impl TripleStore

fn new() -> TripleStore

Constructs a new triple store.

fn count(&self) -> usize

Returns the number of triples that are stored.

fn is_empty(&self) -> bool

Checks if the triple store is empty.

fn add_triple(&mut self, triple: &Triple)

Adds a new triple to the store.

fn remove_triple(&mut self, triple: &Triple)

Deletes the triple from the store.

fn get_triples_with_subject(&self, node: &Node) -> Vec<&Triple>

Returns all triples where the subject node matches the provided node.

fn get_triples_with_predicate(&self, node: &Node) -> Vec<&Triple>

Returns all triples where the predicate node matches the provided node.

fn get_triples_with_object(&self, node: &Node) -> Vec<&Triple>

Returns all triples where the object node matches the provided node.

fn get_triples_with_subject_and_object(&self,
subject_node: &Node,
object_node: &Node)
-> Vec<&Triple>

Returns all triples where the subject and object nodes match the provided nodes.

fn get_triples_with_subject_and_predicate(&self,
subject_node: &Node,
predicate_node: &Node)
-> Vec<&Triple>

Returns all triples where the subject and predicate nodes match the provided nodes.

fn get_triples_with_predicate_and_object(&self,
predicate_node: &Node,
object_node: &Node)
-> Vec<&Triple>

Returns all triples where the predicate and object nodes match the provided nodes.

fn get_blank_nodes(&self) -> Vec<&Node>

Returns all blank nodes of the store.

fn into_vec(self) -> Vec<Triple>

Returns the stored triples as vector.

fn iter(&self) -> Iter<Triple>

Returns an iterator over the stored triples.