Struct rdf_rs::writer::n_triples_writer::NTriplesWriter

pub struct NTriplesWriter { /* fields omitted */ }

RDF writer to generate N-Triples syntax.

Methods

impl NTriplesWriter

fn new() -> NTriplesWriter

Constructor of NTriplesWriter.

Examples

use rdf_rs::writer::n_triples_writer::NTriplesWriter;
use rdf_rs::writer::rdf_writer::RdfWriter;

let writer = NTriplesWriter::new();

fn triple_to_n_triples(&self, triple: &Triple) -> Result<String>

Generates the corresponding N-Triples syntax of the provided triple.

Examples

use rdf_rs::writer::n_triples_writer::NTriplesWriter;
use rdf_rs::writer::rdf_writer::RdfWriter;
use rdf_rs::node::Node;
use rdf_rs::triple::Triple;
use rdf_rs::uri::Uri;

let writer = NTriplesWriter::new();

let subject = Node::BlankNode { id: "blank".to_string() };
let object = Node::LiteralNode { literal: "literal".to_string(), data_type: None, language: Some("en".to_string()) };
let predicate = Node::UriNode { uri: Uri::new("http://example.org/show/localName".to_string()) };
let triple = Triple::new(&subject, &predicate, &object);

assert_eq!(writer.triple_to_n_triples(&triple).unwrap(),
           "_:blank <http://example.org/show/localName> \"literal\"@en .".to_string());

Failures

Invalid node type for a certain position.

fn node_to_n_triples(&self,
node: &Node,
segment: TripleSegment)
-> Result<String>

Converts a single node to its corresponding N-Triples representation.

Checks if the node type is valid considering the triple segment.

Examples

use rdf_rs::writer::n_triples_writer::NTriplesWriter;
use rdf_rs::writer::rdf_writer::RdfWriter;
use rdf_rs::node::Node;
use rdf_rs::triple::TripleSegment;

let writer = NTriplesWriter::new();

let node = Node::BlankNode { id: "blank".to_string() };

assert_eq!(writer.node_to_n_triples(&node, TripleSegment::Subject).unwrap(),
           "_:blank".to_string());

Failures

Node type for triple segment does not conform with NTriples syntax standard.