Crate rdf_rs

This crate provides the tools necessary to parse RDF graphs. It currently contains a full (with very few exceptions) Turtle parser that can parse arbitrary URIs, Triples, and Graphs (see TurtleParser for example usage).

Goals

To provide a simple and easy-to-use RDF parsing API.

To act as an inference engine capable of filling a graph with all the triples that can be inferred from the parsed data.

Usage

This crate is not on crates.io and thus the Cargo.toml entry looks like the following:

[dependencies]
rdf-rs = { git = "https://github.com/CoBrooks/rdf-rs" }

Modules

core

Contains all of the core rdf data structures, such as Uri and Triple.

parsing

Contains the currently-implemented parsers and a base RDFParser trait allowing their creation